+++ packages/angular/src/lib/shared.service.ts
@@ -1,8 +1,10 @@
 import { Injectable } from '@angular/core';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Injectable()
 export class SharedService {
   constructor() {}
+
+  beCool(message: string): string {
+    return `From Library: What's up ${message}`;
+  }
 }