+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ### Features
 
+- angular sharedService DI ([399d2d6](https://github.com/ruslanguns/ruso-monorepo/commit/399d2d683de147525d038253f2252aa6486027ab))
+- base packages for versioning ([b0fc2e2](https://github.com/ruslanguns/ruso-monorepo/commit/b0fc2e276904b690539f7399eabfa42a435f91eb))
+- **all:** monorepo for angular and nestjs ([b83c3ef](https://github.com/ruslanguns/ruso-monorepo/commit/b83c3ef31a5c3a19a33f2e7432657f3ebc7663e2))
+- **module:** angular and nest packages ([ec7705c](https://github.com/ruslanguns/ruso-monorepo/commit/ec7705c0370fcd0d17be69abc2abbb98c61b43a2))
+
+### BREAKING CHANGES
+
+- **module:** build with deps
+
+# 0.0.0 (2020-04-20)
+
+### Features
+
 - base packages for versioning ([b0fc2e2](https://github.com/ruslanguns/ruso-monorepo/commit/b0fc2e276904b690539f7399eabfa42a435f91eb))
 - **all:** monorepo for angular and nestjs ([b83c3ef](https://github.com/ruslanguns/ruso-monorepo/commit/b83c3ef31a5c3a19a33f2e7432657f3ebc7663e2))
 - **module:** angular and nest packages ([ec7705c](https://github.com/ruslanguns/ruso-monorepo/commit/ec7705c0370fcd0d17be69abc2abbb98c61b43a2))